pyproject.toml
[project]
name = "ecouz"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "django-jazzmin>=3.0.1",
    "djangorestframework>=3.16.1",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.29.0",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "redis>=7.1.0",
]


[tool.isort]
profile = "black"
line_length = 88
skip = ".tox, .venv, venv, migrations"
